@@ -5,40 +5,35 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon node_modules/.bin/babel-node index.js",
-    "start": "node_modules/.bin/babel-node index.js",
+    "dev": "nodemon index.js",
+    "start": "node index.js",
     "lint": "./node_modules/.bin/eslint ./ controllers/ class/ --fix"
   },
   "author": "Igor Korsakov <overtorment@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.15.0",
-    "@babel/eslint-parser": "^7.14.2",
-    "@babel/node": "^7.14.9",
-    "@babel/preset-env": "^7.15.6",
-    "@babel/register": "^7.14.5",
-    "@grpc/grpc-js": "^1.3.7",
-    "@grpc/proto-loader": "^0.6.5",
+    "@grpc/grpc-js": "^1.4.1",
+    "@grpc/proto-loader": "^0.6.6",
     "bignumber.js": "^9.0.1",
     "bitcoinjs-lib": "^5.2.0",
     "bolt11": "^1.3.2",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
     "express": "^4.17.1",
-    "express-rate-limit": "^5.3.0",
+    "express-rate-limit": "^5.5.0",
     "frisbee": "^3.1.4",
     "helmet": "^4.6.0",
-    "ioredis": "^4.27.8",
-    "jayson": "^3.6.4",
+    "ioredis": "^4.28.0",
+    "jayson": "^3.6.5",
     "morgan": "^1.10.0",
-    "mustache": "^4.1.0",
-    "prettier": "^2.3.0",
+    "mustache": "^4.2.0",
     "qr-image": "3.2.0",
-    "request": "^2.88.2",
-    "request-promise": "^4.2.6",
     "uuid": "^8.3.2",
     "winston": "^3.3.3"
-  }
+  },
+  "devDependencies": {
+    "eslint": "^8.0.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.4.1"
+  },
+  "type": "module"
 }
