@@ -1,8 +1,9 @@
 // setup lnd rpc
-const config = require('./config');
-var fs = require('fs');
-var grpc = require('@grpc/grpc-js');
-const protoLoader = require('@grpc/proto-loader');
+import config from './config.js';
+import fs from 'fs';
+import grpc from '@grpc/grpc-js';
+import protoLoader from '@grpc/proto-loader';
+
 const loaderOptions = {
   keepCase: true,
   longs: String,
@@ -10,11 +11,12 @@ const loaderOptions = {
   defaults: true,
   oneofs: true,
 };
+
 const packageDefinition = protoLoader.loadSync('rpc.proto', loaderOptions);
-var lnrpc = grpc.loadPackageDefinition(packageDefinition).lnrpc;
+const lnrpc = grpc.loadPackageDefinition(packageDefinition).lnrpc;
 
 process.env.GRPC_SSL_CIPHER_SUITES = 'HIGH+ECDSA';
-var lndCert;
+let lndCert;
 if (process.env.TLSCERT) {
   lndCert = Buffer.from(process.env.TLSCERT, 'hex');
 } else {
@@ -39,7 +41,7 @@ let creds = grpc.credentials.combineChannelCredentials(sslCreds, macaroonCreds);
 // trying to unlock the wallet:
 if (config.lnd.password) {
   process.env.VERBOSE && console.log('trying to unlock the wallet');
-  var walletUnlocker = new lnrpc.WalletUnlocker(config.lnd.url, creds);
+  const walletUnlocker = new lnrpc.WalletUnlocker(config.lnd.url, creds);
   walletUnlocker.unlockWallet(
     {
       wallet_password: Buffer.from(config.lnd.password).toString('base64'),
@@ -54,4 +56,4 @@ if (config.lnd.password) {
   );
 }
 
-module.exports = new lnrpc.Lightning(config.lnd.url, creds, { 'grpc.max_receive_message_length': 1024 * 1024 * 1024 });
+export default new lnrpc.Lightning(config.lnd.url, creds, { 'grpc.max_receive_message_length': 1024 * 1024 * 1024 });
